+++ src/movie/models/movie_model.py
@@ -1,6 +1,6 @@
 class Movie:
-    def __init__(self, movie_id, title, genre, duration, release_date):
-        self.id = movie_id
+    def __init__(self, id, title, genre, duration, release_date):
+        self.id = id
         self.title = title
         self.genre = genre
         self.duration = duration
@@ -13,4 +13,17 @@ class Movie:
             "genre": self.genre,
             "duration": self.duration,
             "release_date": self.release_date
+        }
+
+class Showtime:
+    def __init__(self, id, movie_id, start_time):
+        self.id = id
+        self.movie_id = movie_id
+        self.start_time = start_time
+    
+    def to_dict(self):
+        return {
+            "id": self.id,
+            "movie_id": self.movie_id,
+            "start_time": self.start_time
         }