+++ src/movie/persistence/movie_repository.py
@@ -1,20 +1,43 @@
+import sqlite3
+import os
 from movie.models.movie_model import Movie
 
-# In-memory storage for now
-movie_db = {}
-next_id = 1
-
 class MovieRepository:
-    def save(self, title, genre, duration, release_date):
-        global next_id
-        movie_id = str(next_id)
-        new_movie = Movie(movie_id, title, genre, duration, release_date)
-        movie_db[movie_id] = new_movie
-        next_id += 1
-        return new_movie
+    def __init__(self):
+        current_dir = os.path.dirname(os.path.abspath(__file__)) 
+        self.db_path = os.path.join(current_dir, '..', 'db', 'movies.db')
 
-    def find_by_id(self, movie_id):
-        return movie_db.get(movie_id)
+        with sqlite3.connect(self.db_path) as conn:
+            conn.execute("""CREATE TABLE IF NOT EXISTS movies (
+                            id INTEGER PRIMARY KEY AUTOINCREMENT,
+                            title TEXT NOT NULL,
+                            genre TEXT,
+                            duration INTEGER NOT NULL,
+                            release_date TEXT)""")
+            
+            conn.execute("""CREATE TABLE IF NOT EXISTS showtimes (
+                            id INTEGER PRIMARY KEY AUTOINCREMENT,
+                            movie_id INTEGER NOT NULL,
+                            start_time TEXT NOT NULL)""")
 
-    def find_all(self):
-        return list(movie_db.values())
+    def add_movie(self, title, genre, duration, release_date):
+        with sqlite3.connect(self.db_path) as conn: # Dùng self.db_path
+            cursor = conn.cursor()
+            sql = "INSERT INTO movies (title, genre, duration, release_date) VALUES (?, ?, ?, ?)"
+            cursor.execute(sql, (title, genre, duration, release_date))
+            return cursor.lastrowid
+
+    def get_movie(self, movie_id):
+        with sqlite3.connect(self.db_path) as conn:
+            cursor = conn.cursor()
+            cursor.execute("SELECT * FROM movies WHERE id = ?", (movie_id,))
+            row = cursor.fetchone()
+            if row:
+                return Movie(id=row[0], title=row[1], genre=row[2], duration=row[3], release_date=row[4])
+            return None
+
+    def add_showtime(self, movie_id, start_time):
+        with sqlite3.connect(self.db_path) as conn:
+            cursor = conn.cursor()
+            cursor.execute("INSERT INTO showtimes (movie_id, start_time) VALUES (?, ?)", (movie_id, start_time))
+            return cursor.lastrowid